requirements.txt
Flask
pillow
pymongo
pyyaml
imagehash